fix(migration): use FirstOrCreate instead of clause.OnConflict
Since we are using one upserted object in the next query, it is imperative that the upserted object return the ID of the existing object on conflict so that we can use it in the following query. However, since clause.OnConflict will literally do nothing on conflict, the ID field of godRight remained unset. This causes a fk error when the next upsert query tries to insert just the role_id in role_right table, leading the value of right_id for that insert query to be set to some random UUID which mostly likely not be present in the rights table.

Signed-off-by: ksdfg <ksdfg123@gmail.com>